sfef

+++ src/carlorolf/Grass.java
@@ -5,7 +5,7 @@ import java.util.Random;
 public class Grass extends VisibleObject
 {
     private Random random;
-    public Grass(int x,int y){
+    public Grass(double x, double y){
 	super(x, y, 1, 1, Images.getImage("grass0.png"));
 	random = new Random();
     }
